Gram Schmidt Process
The Gram-Schmidt process is a method for constructing an orthonormal basis of a space that a set of given vectors span. It can also be used to determine the dimension of that space, which may be different than the dimension of the vectors themselves or the number of vectors provided to span the space.
This notebook computes a Gram-Schmidt basis, and gives the dimension of a span of vectors.