Identifying special matrices
When coding or solving data analysis problems, one problem that can occur is if your code encounters a special matrix that isn't invertible, or has an infinite number of eigenvectors, or similar. On other occasions, for example where you are reducing dimensionality, that might even be desirable! So here you will write a code fragment that traps for different types of special matrices before calling the python inversion routine, and classifies the type of special case encountered.